fix: Link to registration on SelectServer page
